fix: fix type errors and application.properties

+++ src/main/duerre-manager/src/model/editor/managers/state-manager.ts
@@ -15,6 +15,7 @@ import { Line } from "../core/shape/line";
 import { BezierLine } from "../core/shape/bezier-line";
 import { toVec2DArray, vec2DEquals } from "../core/math/vec2d";
 import { toast } from "vue3-toastify";
+import Client from '@/plugins/http/openapi';
 
 export class StateManager extends GenericManager {
 
@@ -132,12 +133,14 @@ export class StateManager extends GenericManager {
         return { lines, valid };
     }
 
-    public load(die: IDieShapeImport): boolean {
+    public load(die: IDieShapeImport | Client.Components.Schemas.DieData): boolean {
+        if(!die.lines) return false;
+
         this.editor.clear();
 
         const points: Point[] = [];
         die.lines.forEach(l => {
-            const vecs = toVec2DArray(l.points);
+            const vecs = toVec2DArray(l.points!);
             vecs.forEach(v => {
                 const index = points.findIndex(p => p.equalsByVector(v))
                 if (index == -1) {
@@ -149,11 +152,11 @@ export class StateManager extends GenericManager {
 
         die.lines.forEach(l => {
             if (l.type == 'line' || l.type == 'bezier') {
-                const vecs = toVec2DArray(l.points);
+                const vecs = toVec2DArray(l.points!);
                 const shapePoints: Point[] = vecs.map(v => points.find(p => p.equalsByVector(v))!);
                 console.log(shapePoints);
                 const drawingLine = new MeasurableShape<Line | BezierLine>(this.editor, { x: 0, y: 0 }, l.type == 'line' ? Line : BezierLine);
-                drawingLine.updatePoints(l.points);
+                drawingLine.updatePoints(l.points!);
                 drawingLine.extShape.overrideStartPoint(shapePoints[0]);
                 drawingLine.extShape.overrideEndPoint(shapePoints[2] ?? shapePoints[1]);
                 this.editor.layer.add(drawingLine.group);
@@ -164,7 +167,7 @@ export class StateManager extends GenericManager {
         let cutLineLoadError = false;
         die.lines.forEach(l => {
             if (l.type == 'cut') {
-                const [startPoint, endPoint] = toVec2DArray(l.points);
+                const [startPoint, endPoint] = toVec2DArray(l.points!);
                 const startPointShape = this.findNearestVertex(startPoint);
                 const endPointShape = this.findNearestVertex(endPoint);
                 if (!startPointShape?.shape || !endPointShape?.shape) {
